requirements.txt
Flask
line-bot-sdk==2.4.1
gunicorn
